add RemoveClientsViewController
