+++ wire-ios/Wire-iOS/Sources/UserInterface/Settings/DeviceManagement/ClientTableViewCell.swift
@@ -127,85 +127,3 @@ final class ClientTableViewCell: UITableViewCell {
         super.prepareForReuse()
     }
 }
-
-final class ClientTableViewCell1: UITableViewCell {
-
-    typealias LabelColors = SemanticColors.Label
-
-    // MARK: - Properties
-
-    private let deviceNameLabel = DynamicFontLabel(
-        style: .headline,
-        color: LabelColors.textDefault)
-
-    private let proteusIdLabel = DynamicFontLabel(
-        style: .caption1,
-        color: LabelColors.textCellSubtitle)
-
-    private let activationDateLabel = DynamicFontLabel(
-        style: .caption1,
-        color: LabelColors.textCellSubtitle)
-
-    var viewModel: ClientTableViewCellModel? {
-        didSet {
-            deviceNameLabel.text = viewModel?.title
-            if let date = viewModel?.activationDate?.formattedDate {
-                activationDateLabel.text = L10n.Localizable.Registration.Devices.activated(date)
-            }
-            proteusIdLabel.text = viewModel?.proteusLabelText
-        }
-    }
-
-    // MARK: - Initialization
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-
-        createConstraints()
-        setupStyle()
-    }
-
-    @available(*, unavailable)
-    required init(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - Methods
-
-    private func setupStyle() {
-        deviceNameLabel.accessibilityIdentifier = "device name"
-        proteusIdLabel.accessibilityIdentifier = "device proteus ID"
-        activationDateLabel.accessibilityIdentifier = "activation date"
-
-        backgroundColor = SemanticColors.View.backgroundUserCell
-
-        addBorder(for: .bottom)
-    }
-
-    private func createConstraints() {
-        [
-            deviceNameLabel,
-            proteusIdLabel,
-            activationDateLabel
-        ].forEach { view in
-            view.translatesAutoresizingMaskIntoConstraints = false
-            contentView.addSubview(view)
-        }
-
-        NSLayoutConstraint.activate([
-            deviceNameLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 16),
-            deviceNameLabel.leftAnchor.constraint(equalTo: contentView.leftAnchor, constant: 16),
-            deviceNameLabel.rightAnchor.constraint(lessThanOrEqualTo: contentView.rightAnchor, constant: -16),
-
-            proteusIdLabel.topAnchor.constraint(equalTo: deviceNameLabel.bottomAnchor, constant: 10),
-            proteusIdLabel.leftAnchor.constraint(equalTo: contentView.leftAnchor, constant: 16),
-            proteusIdLabel.rightAnchor.constraint(lessThanOrEqualTo: contentView.rightAnchor, constant: -16),
-
-            activationDateLabel.topAnchor.constraint(equalTo: proteusIdLabel.bottomAnchor, constant: 10),
-            activationDateLabel.leftAnchor.constraint(equalTo: contentView.leftAnchor, constant: 16),
-            activationDateLabel.rightAnchor.constraint(lessThanOrEqualTo: contentView.rightAnchor, constant: -16),
-            activationDateLabel.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -16)
-        ])
-    }
-
- }
